abstract out splitTime
