+++ src/coreUtils/splitTime.js
@@ -1,0 +1,14 @@
+export function splitTime(s) {
+  const ms = s % 1000;
+  s = (s - ms) / 1000;
+  const secs = s % 60;
+  s = (s - secs) / 60;
+  const mins = s % 60;
+  const hrs = (s - mins) / 60;
+  return [
+    hrs,
+    mins,
+    secs,
+    ms,
+  ];
+}
